requirements.txt
streamlit~=1.28.1
pandas~=2.1.0
numpy~=1.26.0
validators~=0.22.0
scikit-learn~=1.3.1
fuzzywuzzy~=0.18.0
yadisk~=1.3.3
PyYAML~=6.0.1